# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PhotoSet 摄影套图浏览平台后端 API 文档\n提供套图管理、社区互动、用户认证、支付等功能",
        "title": "PhotoSet API",
        "termsOfService": "https://photoset.example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@photoset.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/auth/captcha": {
            "get": {
                "description": "生成登录/注册/找回密码用的图形验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "生成图形验证码",
                "parameters": [
                    {
                        "type": "string",
                        "default": "login",
                        "description": "场景(login/register/forgot)",
                        "name": "action",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码ID和Base64图片",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "生成失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "在请求头中添加 Bearer {token} 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}